+++ gems/aws-sdk-core/lib/aws-sdk-sts/plugins/sts_regional_endpoints.rb
@@ -5,12 +5,12 @@ module Aws
       class STSRegionalEndpoints < Seahorse::Client::Plugin
 
         option(:sts_regional_endpoints,
-          default: 'legacy',
+          default: 'regional',
           doc_type: String,
           docstring: <<-DOCS) do |cfg|
 Passing in 'regional' to enable regional endpoint for STS for all supported
-regions (except 'aws-global'), defaults to 'legacy' mode, using global endpoint
-for legacy regions.
+regions (except 'aws-global'). Using 'legacy' mode will force all legacy 
+regions to resolve to the STS global endpoint.
           DOCS
           resolve_sts_regional_endpoints(cfg)
         end
@@ -22,7 +22,7 @@ for legacy regions.
           env_mode = nil if env_mode == ''
           cfg_mode = Aws.shared_config.sts_regional_endpoints(
             profile: cfg.profile)
-          env_mode || cfg_mode || 'legacy'
+          env_mode || cfg_mode || 'regional'
         end
 
       end