+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Issue - Change the default of `sts_regional_endpoints` from 'legacy' to 'regional'.
+
 3.91.1 (2020-03-10)
 ------------------